% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_coverage.R
\name{cha_coverage}
\alias{cha_coverage}
\title{List Topic Coverage}
\usage{
cha_coverage(topic_key, progress = TRUE)
}
\arguments{
\item{topic_key}{3-8 letter ID uniquely identifying
a topic.}
}
\value{
Topic coverage tibble
}
\description{
List all combinations of population, periods, and
geographic layers available for a given topic. To
search for individual topics use \code{cha_topics}.
}
\examples{
cha_coverage("POP")
}
