% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_data.R
\name{cha_data}
\alias{cha_data}
\title{Obtain Chicago Health Atlas Data}
\usage{
cha_data(topic_key, population_key, period_key, layer_key)
}
\arguments{
\item{topic_key}{Unique ID specifying a topic.}

\item{population_key}{Unique ID for a population
stratification.}

\item{period_key}{Unique ID for a time period.}

\item{layer_key}{Unique ID for a geographic layer.}
}
\value{
Data tibble containing value and standard
error for topic measure.
}
\description{
Obtain data for a topic within a population at a
specified time period and a geographic scale.
}
\examples{
cha_data("POP", "H", "2014-2018", "zip")
}
