% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_layer.R
\name{cha_layer}
\alias{cha_layer}
\title{Obtain Geographic Layer}
\usage{
cha_layer(layer_key, progress = TRUE)
}
\arguments{
\item{layer_key}{Unique ID for a geographic layer.}

\item{progress}{Display a progress bar?}
}
\value{
\code{sf} geographic layer.
}
\description{
Import geographic layer as a \code{sf} object.
}
\examples{
cha_layer("zip")
}
