% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_topic_subcategories.R
\name{cha_topic_subcategories}
\alias{cha_topic_subcategories}
\title{List Topic Subcategories}
\usage{
cha_topic_subcategories()
}
\value{
Topic subcategory tibble
}
\description{
List Topic Subcategories
}
\examples{
cha_topic_subcategories()
}
