% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha_topics.R
\name{cha_topics}
\alias{cha_topics}
\title{List Population Stratifications}
\usage{
cha_topics(subcategory_key = NULL, progress = TRUE)
}
\arguments{
\item{subcategory_key}{Unique id for a
topic subcategory}

\item{progress}{Display a progress bar?}
}
\value{
Topics information tibble
}
\description{
List Population Stratifications
}
\examples{
cha_topics("education", progress = FALSE)
}
