% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_coverage.R
\name{ha_coverage}
\alias{ha_coverage}
\title{List Topic Coverage}
\usage{
ha_coverage(topic_key, progress = TRUE)
}
\arguments{
\item{topic_key}{3-8 letter ID uniquely identifying
a topic.}

\item{progress}{Display a progress bar?}
}
\value{
Topic coverage tibble
}
\description{
List all combinations of population, periods, and
geographic layers available for a given topic. To
search for individual topics use \code{ha_topics}.
}
\examples{
ha_set("chicagohealthatlas.org")

ha_coverage("POP", progress = FALSE)
}
