% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_layer.R
\name{ha_layer}
\alias{ha_layer}
\title{Obtain Geographic Layer}
\usage{
ha_layer(layer_key, progress = TRUE)
}
\arguments{
\item{layer_key}{Unique ID for a geographic layer.}

\item{progress}{Display a progress bar?}
}
\value{
\code{sf} geographic layer.
}
\description{
Import geographic layer as a \code{sf} object.
}
\examples{
\donttest{
ha_set("chicagohealthatlas.org")

ha_layer("zip", progress = FALSE)
}
}
