% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_layer.R
\name{ha_layers}
\alias{ha_layers}
\title{List Geographic Layers}
\usage{
ha_layers()
}
\value{
Layer information tibble.
}
\description{
List all geographic layers available.
}
\examples{
\donttest{
ha_set("chicagohealthatlas.org")

ha_layers()
}
}
