% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_point_layer.R
\name{ha_point_layer}
\alias{ha_point_layer}
\title{Obtain Point Layer}
\usage{
ha_point_layer(point_layer_uuid)
}
\arguments{
\item{point_layer_uuid}{Unique ID for a point layer.}
}
\value{
\code{sf} point layer.
}
\description{
Import point layer as a \code{sf} object.
}
\examples{
\donttest{
ha_set("chicagohealthatlas.org")

ha_point_layer("67f58fa0-0dfa-4ee9-8600-c1ab4a093dc6")
}
}
