% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_point_layer.R
\name{ha_point_layers}
\alias{ha_point_layers}
\title{List Point Layers}
\usage{
ha_point_layers()
}
\value{
Point layer information tibble.
}
\description{
List all point layers available.
}
\examples{
\donttest{
ha_set("chicagohealthatlas.org")

ha_point_layers()
}
}
