% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_topic_subcategories.R
\name{ha_topic_subcategories}
\alias{ha_topic_subcategories}
\title{List Topic Subcategories}
\usage{
ha_topic_subcategories()
}
\value{
Topic subcategory tibble
}
\description{
List the topic subcategories, which can be
used to filter topics within \code{ha_topics()}.
}
\examples{
ha_set("chicagohealthatlas.org")

ha_topic_subcategories()
}
