% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_topics.R
\name{ha_topics}
\alias{ha_topics}
\title{List Topics}
\usage{
ha_topics(subcategory_key = NULL, progress = TRUE)
}
\arguments{
\item{subcategory_key}{Unique id for a
topic subcategory}

\item{progress}{Display a progress bar?}
}
\value{
Topics information tibble.
}
\description{
List all topics available with descriptions
and units.
}
\examples{
\donttest{
ha_set("chicagohealthatlas.org")

ha_topics("education", progress = FALSE)
}
}
