% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_coverage.R
\name{ha_coverage}
\alias{ha_coverage}
\title{List Topic Coverage}
\usage{
ha_coverage(topic_key, layer_key = NULL, keys_only = FALSE, progress = TRUE)
}
\arguments{
\item{topic_key}{Unique ID specifying a topic.}

\item{layer_key}{Character string or vector of
Unique IDs for geographic layers.}

\item{keys_only}{Return only keys?}

\item{progress}{Display a progress bar?}
}
\value{
Topic coverage tibble.
}
\description{
List all combinations of population, periods, and
geographic layers available for a given topic. To
search for individual topics use \code{ha_topics()}.
}
\examples{
\donttest{
ha_set("chicagohealthatlas.org")

ha_coverage("POP", progress = FALSE)
}
}
