% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_data.R
\name{ha_data}
\alias{ha_data}
\title{Obtain Data}
\usage{
ha_data(topic_key, population_key, period_key, layer_key, wide = FALSE)
}
\arguments{
\item{topic_key}{Unique ID specifying a topic.}

\item{population_key}{Unique ID for a population
stratification.}

\item{period_key}{Unique ID for a time period.}

\item{layer_key}{Unique ID for a geographic layer.}

\item{wide}{Pivot wider so that each row contains
the values of all topics within a population at a
specified time period and a geographic area?}
}
\value{
Data tibble containing value and standard
error for each topic measure.
}
\description{
Obtain data for one or more topics within specificied
populations for specified time periods and geographic
scale.

All combinations of topic, population, period, and layer
will be returned. Warnings will be generated for any
invalid combinations.
}
\examples{
ha_set("chicagohealthatlas.org")

ha_data("POP", "H", "2014-2018", "zip")
}
