% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_getset.R
\name{ha_set}
\alias{ha_set}
\title{Set Health Atlas Portal}
\usage{
ha_set(ha_URL)
}
\arguments{
\item{ha_URL}{URL of the health atlas home page.}
}
\value{
No return value. Sets ha_URL environment variable.
}
\description{
Set health atlas to connect to.
}
\examples{
\donttest{
ha_set("chicagohealthatlas.org")
}
}
